+++ demos/vue-nestjs/vite.config.ts
@@ -10,6 +10,9 @@ export default defineConfig({
         target: 'http://localhost:4002',
         ws: true,
       },
+      '/datasole-worker.iife.min.js': {
+        target: 'http://localhost:4002',
+      },
     },
   },
   build: {